fix(appCache): add Fallback, fixes Firefox trouble
Firefox could not reload the page after it was loaded initially. Adding the FALLBACK section
fixes that. It is not optimal as any URL that cannot be found will now return /index.html,
including assets like CSS, JS and image files. But that should not be a problem for this app.